+++ packages/datasource-toolkit/src/interfaces/action.ts
@@ -151,6 +151,17 @@ export type ActionFieldNumberInputList = ActionFieldBase & {
   step?: number;
 };
 
+export type ActionFieldCurrencyInput = ActionFieldBase & {
+  type: 'Number';
+  widget: 'CurrencyInput';
+  placeholder?: string;
+  min?: number;
+  max?: number;
+  step?: number;
+  currency: string;
+  base?: 'Unit' | 'Cent';
+};
+
 export type ActionFieldDropdownAll =
   | ActionFieldDropdown<'Date' | 'Dateonly' | 'Number' | 'String' | 'StringList', string>
   | ActionFieldDropdown<'Number' | 'NumberList', number>;
@@ -180,6 +191,7 @@ export type ActionField = StrictUnion<
   | ActionFieldColorPicker
   | ActionFieldNumberInputList
   | ActionFieldDateInput
+  | ActionFieldCurrencyInput
 >;
 
 export type ActionFieldWidget =